Improve login/register layout
Implemented an activity called BackgroundImageActivity that extends BOActivity and will display its layout in a responsive manner. If the screen is too large to let the layout fill the whole width of the screen, the width of the layout will be set to a fixed value. Furthermore, it is now possible to scroll the login/register screen. The shadows of the buttons have been removed and the keyboard will be kept from opening when the activity starts.