Add custom BOFlatButton that is able to show a loading indicator
The BOFlatButton can now be used to display buttons that are either orange or white. All settings of style, text size etc. will be done automatically and in a consistent way for all these Buttons. It is also possible to call setShowLoadingIndicator(true) to display an indeterminately. The custom BOEditTexts have been applied to the LoginRegisterActivity.