Restructure ProfileFragment, add check for user data changes and some UI changes
The ProfileFragment now has a title, a back arrow (instead of burger icon) and is able to display a loading indicator in the toolbar. It will check for changes in the user data and ask to save changes before quitting. The fragment has been restructured and the dividers are now lighter. Fixed a bug when updating a user with empty participant values (user would become participant before). Added autofill for become participant view when data has already been provided.